move merge logic to utils directory
